app/src/main/java/be/acuzio/mrta/LocalBroadcastActivity.java
package be.acuzio.mrta;

import android.app.Activity;
import android.content.BroadcastReceiver;
import android.content.Context;
import android.content.Intent;
import android.content.IntentFilter;
import android.os.Bundle;
import android.support.v4.content.LocalBroadcastManager;
import android.widget.Toast;

import be.acuzio.mrta.util.Logger;

public class LocalBroadcastActivity extends Activity {
    final private BroadcastReceiver myReceiver = new BroadcastReceiver() {
        @Override
        public void onReceive(Context context, Intent intent) {
            Logger.d(TAG, "onReceive called");

            Toast.makeText(context, "onReceive Toast Called", Toast.LENGTH_SHORT).show();
        }
    };

    private final static String TAG = LocalBroadcastActivity.class.getSimpleName();

    @Override
    protected void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);
        setContentView(R.layout.activity_local_broadcast);
    }

    @Override
    protected void onPause() {
        super.onPause();
        LocalBroadcastManager.getInstance(this).unregisterReceiver(this.myReceiver);
    }

    @Override
    protected void onResume() {
        super.onResume();

        LocalBroadcastManager.getInstance(this).registerReceiver(myReceiver, new IntentFilter("be.acuzio.mrta.LOCAL_BROADCAST"));
    }
}
